seems strange to mark this warning as boring
Probably this used to be marked as boring because we didn't have a good
way to filter out warnings that we knew were uninteresting after they
were marked as warnings. But now we do, so let's go back to making it a
warning, and let the user say which warnings he wants to ignore.